Roll multiple binaries into a single cli with sub-commands
